src/main/java/tfw/immutable/iis/intiis/IntIisFromIntIla.java
package tfw.immutable.iis.intiis;

import java.io.IOException;
import tfw.check.Argument;
import tfw.immutable.iis.IisFromIlaUtil;
import tfw.immutable.ila.intila.IntIla;

public final class IntIisFromIntIla {
    private IntIisFromIntIla() {}

    public static IntIis create(final IntIla ila) {
        return new IntIisImpl(ila);
    }

    private static class IntIisImpl extends AbstractIntIis {
        private final IntIla ila;

        private long index = 0;

        public IntIisImpl(final IntIla ila) {
            Argument.assertNotNull(ila, "ila");

            this.ila = ila;
        }

        @Override
        public void closeImpl() throws IOException {
            index = ila.length();
        }

        @Override
        protected int readImpl(int[] array, int offset, int length) throws IOException {
            final int elementsToGet = IisFromIlaUtil.read(ila.length(), index, length);

            if (elementsToGet > -1) {
                ila.get(array, offset, index, elementsToGet);

                index += elementsToGet;
            }

            return elementsToGet;
        }

        @Override
        protected long skipImpl(long n) throws IOException {
            final long elementsSkipped = IisFromIlaUtil.skip(ila.length(), index, n);

            if (elementsSkipped > -1) {
                index += elementsSkipped;
            }

            return elementsSkipped;
        }
    }
}
// AUTO GENERATED FROM TEMPLATE
